+++ python/cykooz_heif/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 :Authors: cykooz
 :Date: 23.06.2019