+++ python/cykooz_heif/image.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 :Authors: cykooz
 :Date: 25.06.2019